+++ Core/Application/Interfaces/Repositories/IMicroRegionRepository.cs
@@ -5,6 +5,6 @@ namespace Application.Interfaces.Repositories;
 public interface IMicroRegionRepository : IRepository<MicroRegion>
 {
     Task<List<MicroRegion>> GetAllAsync();
-    Task<MicroRegion> GetByIdAsync(int id);
-    Task<List<MicroRegion>> GetByMesoregionIdAsync(int mesoregionId);
+    Task<MicroRegion> GetByIdAsync(long id);
+    Task<List<MicroRegion>> GetByMesoregionIdAsync(long mesoregionId);
 }
