+++ Core/Application/Interfaces/Repositories/IMunicipalityRepository.cs
@@ -5,6 +5,6 @@ namespace Application.Interfaces.Repositories;
 public interface IMunicipalityRepository : IRepository<Municipality>
 {
     Task<List<Municipality>> GetAllAsync();
-    Task<Municipality> GetByIdAsync(int id);
-    Task<List<Municipality>> GetByMicroRegionIdAsync(int microRegionId);
+    Task<Municipality> GetByIdAsync(long id);
+    Task<List<Municipality>> GetByMicroRegionIdAsync(long microRegionId);
 }
