+++ Core/Application/Interfaces/Repositories/IStateRepository.cs
@@ -5,5 +5,5 @@ namespace Application.Interfaces.Repositories;
 public interface IStateRepository : IRepository<State>
 {
     Task<List<State>> GetAllAsync();
-    Task<State> GetByIdAsync(int id);
+    Task<State> GetByIdAsync(long id);
 }
