+++ Infraestructure/Elasticsearch/DTOs/StateDto.cs
@@ -2,8 +2,8 @@
 
 public class StateDto
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Name { get; set; } = null!;
     public string Initials { get; set; } = null!;
-    public int RegionId { get; set; }
+    public long RegionId { get; set; }
 }
