+++ Infraestructure/Elasticsearch/DTOs/SubDistrictsDto.cs
@@ -2,7 +2,7 @@
 
 public class SubDistrictsDto
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Name { get; set; } = null!;
-    public int DistrictId { get; set; }
+    public long DistrictId { get; set; }
 }
